Added the portfolio page
Added a portfolio page that contains cards including information about what I have done, and includes a way to visit the github pages for the projects. And it also has a link to the github repo. It would be nice to one day make a API that would return this information and then use JS to dynamically create this content. But we can hand-jam it in for now. 😀